Updated to depend on pydicom 0.9.7 and use DICOM standard keywords for meta data.
DICOM elments with VR of DS and IS are converted to float and int while extracting.
Filtering of meta data is done by DicomStack object just before converting to Nifti.
Elements is VR of OB, OW, or UN are only ignored if the value contains non-ascii bytes.
The dcmstack cli is refactored to reflect these changes.